menu admin funcional hasta terminal
voy subiendo esto por que estaba terminando de hacer funcional hasta trenes y el de agregar trenes me rompio todo el programa, mejor prevenir que curar

+++ src/tren/Tren.java
@@ -14,7 +14,7 @@ public class Tren implements Serializable{
 		///
 	}
 
-	public Tren(String modelo, int anioFabricacion, int distanciaMaxima, int precioPorKilometro){
+	public Tren(String modelo, int anioFabricacion, double distanciaMaxima, int precioPorKilometro){
 		this.precioPorKilometro = precioPorKilometro;
 		this.anioFabricacion = anioFabricacion;
 		this.modelo = modelo;
